+++ test/test.js
@@ -1,4 +1,4 @@
-"use strict";
+'use strict';
 
 var SG = require('..');
 
@@ -13,8 +13,7 @@ describe('Savitzky–Golay test', function () {
         var data = new Array(200);
         for (var i = 0; i < data.length; i++)
             data[i] = Math.sin(i);
-        var model = new SG(data,1);
-        var ans = model.calc(options);
+        var ans = new SG(data,1, options);
         for (var j = 4; j < ans.length - 4; j++)
             ans[j].should.be.approximately(data[j], 0.08);
     });
@@ -28,8 +27,7 @@ describe('Savitzky–Golay test', function () {
         var data = new Array(200);
         for (var i = 0; i < data.length; i++)
             data[i] = Math.sin(i);
-        var model = new SG(data,1);
-        var ans = model.calc(options);
+        var ans = new SG(data,1, options);
         for (var j = 4; j < ans.length - 4; j++)
             ans[j].should.be.approximately(Math.cos(j), 0.05);
     });