+++ History.md
@@ -3,6 +3,7 @@ HEAD
 
 * change numeric.js for ml-matrix
 * change variable and option names
+* remove the instance and just have a function
 
 1.0.0
 ============